# HearingsApi/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "HearingsApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Hearings/Get": {
      "get": {
        "tags": [
          "Hearings"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Hearings"
              }
            }
          }
        }
      }
    },
    "/Hearings/GetFromDB": {
      "get": {
        "tags": [
          "Hearings"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Hearings"
              }
            }
          }
        }
      }
    },
    "/Hearings/{processNumber}": {
      "get": {
        "tags": [
          "Hearings"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "processNumber",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Hearings"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "processNumber",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Hearings"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Hearings"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "processNumber",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Hearings/AddToDB": {
      "post": {
        "tags": [
          "Hearings"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Hearings"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Hearings": {
      "type": "object",
      "properties": {
        "processNumber": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "court": {
          "type": "string"
        },
        "correspondent": {
          "type": "string"
        }
      }
    }
  }
}